# .scratch/source/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Casino Admin Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /telegram-buttons:
    get:
      operationId: listTelegramButtons
      tags: [telegram-buttons]
      summary: List Telegram buttons
      responses:
        "200":
          description: List of buttons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelegramButton"
    post:
      operationId: createTelegramButton
      tags: [telegram-buttons]
      summary: Create a Telegram button
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramButtonInput"
      responses:
        "201":
          description: Created button
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramButton"
  /telegram-buttons/{id}:
    patch:
      operationId: updateTelegramButton
      tags: [telegram-buttons]
      summary: Update a Telegram button
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramButtonUpdate"
      responses:
        "200":
          description: Updated button
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramButton"
    delete:
      operationId: deleteTelegramButton
      tags: [telegram-buttons]
      summary: Delete a Telegram button
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── SETTINGS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    TelegramButton:
      type: object
      properties:
        id:
          type: number
        label:
          type: string
        action:
          type: string
        order:
          type: number
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, label, action, order, enabled, createdAt, updatedAt]
    TelegramButtonInput:
      type: object
      properties:
        label:
          type: string
        action:
          type: string
        order:
          type: number
      required: [label, action]
    TelegramButtonUpdate:
      type: object
      properties:
        label:
          type: string
        action:
          type: string
        order:
          type: number
        enabled:
          type: boolean

    # ─── SETTINGS ─────────────────────────────────────────────────────────────────
...