# .scratch/source/lib/api-zod/src/generated/types/telegramButtonInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Casino Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramButtonInput {
  label: string;
  action: string;
  order?: number;
}